@@ -3,7 +3,7 @@ import numpy as np
 
 import triangle as tr
 
-pts = np.array([[0, 0], [0, 1], [.5, .5], [1, 1], [1, 0]])
+pts = np.array([[0, 0], [0, 1], [0.5, 0.5], [1, 1], [1, 0]])
 tri = tr.delaunay(pts)
 
 A = dict(vertices=pts)
