@@ -2,10 +2,10 @@ import matplotlib.pyplot as plt
 
 import triangle as tr
 
-spiral = tr.get_data('spiral')
+spiral = tr.get_data("spiral")
 a = tr.triangulate(spiral)
-b = tr.triangulate(spiral, 'q')
-c = tr.triangulate(spiral, 'q32.5')
+b = tr.triangulate(spiral, "q")
+c = tr.triangulate(spiral, "q32.5")
 
 plt.figure(figsize=(5, 6))
 ax1 = plt.subplot(221)
