@@ -1,18 +1,4 @@
+from .data import get_data, load, loads, show_data
+from .plot import compare, comparev, plot
+from .tri import convex_hull, delaunay, triangulate, voronoi
 from .version import __version__
-from .data import (
-    loads,
-    load,
-    get_data,
-    show_data,
-)
-from .plot import (
-    plot,
-    comparev,
-    compare,
-)
-from .tri import (
-    triangulate,
-    convex_hull,
-    delaunay,
-    voronoi,
-)
